"Add notification microservice with protobuf files and scripts
This commit introduces a new notification microservice.
It includes a script to generate Go protobufs for message passing,
as well as the generated Go code for the notification.proto and
notification_service.proto files.
These files will facilitate the creation of notifications,
registration of devices for notifications, and unregistration for the same."